# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: DEVLINK API
  description: API service for the Distributed Link Intelligence Platform.
  version: 0.1.0
paths:
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register a user
      description: Create a password-backed user account using a normalized email.
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Register Api V1 Auth Register Post
        '401':
          description: Invalid or expired credentials.
        '409':
          description: The requested account already exists.
        '422':
          description: Request validation failed.
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Log in
      description: Verify password credentials and issue JWT access/refresh tokens.
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired credentials.
        '409':
          description: The requested account already exists.
        '422':
          description: Request validation failed.
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh tokens
      description: Rotate a refresh token and return a new access/refresh token pair.
        The previous refresh token is revoked.
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired credentials.
        '409':
          description: The requested account already exists.
        '422':
          description: Request validation failed.
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Log out
      description: Revoke a refresh token so it can no longer be rotated.
      operationId: logout_api_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '401':
          description: Invalid or expired credentials.
        '409':
          description: The requested account already exists.
        '422':
          description: Request validation failed.
  /api/v1/auth/oauth/google:
    get:
      tags:
      - auth
      summary: Start Google OAuth
      description: Placeholder endpoint for a future Google OAuth authorization flow.
      operationId: start_google_oauth_api_v1_auth_oauth_google_get
      responses:
        '501':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Start Google Oauth Api V1 Auth Oauth Google Get
        '401':
          description: Invalid or expired credentials.
        '409':
          description: The requested account already exists.
        '422':
          description: Request validation failed.
  /api/v1/auth/oauth/google/callback:
    get:
      tags:
      - auth
      summary: Finish Google OAuth
      description: Placeholder callback endpoint for a future Google OAuth flow.
      operationId: finish_google_oauth_api_v1_auth_oauth_google_callback_get
      responses:
        '501':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Finish Google Oauth Api V1 Auth Oauth Google Callback
                  Get
        '401':
          description: Invalid or expired credentials.
        '409':
          description: The requested account already exists.
        '422':
          description: Request validation failed.
...
  /api/v1/analytics/{link_id}/summary:
    get:
      tags:
      - analytics
      summary: Get analytics summary
      description: Return total clicks for a link using the denormalized link counter.
        Unique clicks are not tracked yet and are reported as unavailable.
      operationId: get_summary_api_v1_analytics__link_id__summary_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
          title: Link Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link.
        '404':
          description: Link not found.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
  /api/v1/analytics/{link_id}/timeseries:
    get:
      tags:
      - analytics
      summary: Get click time series
      description: Return daily click counts for a link from the aggregate analytics
        table. Dashboard queries never read raw click events.
      operationId: get_timeseries_api_v1_analytics__link_id__timeseries_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
          title: Link Id
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
          title: From
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
          title: To
      - name: granularity
        in: query
        required: false
        schema:
          const: day
          type: string
          default: day
          title: Granularity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeseriesResponse'
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link.
        '404':
          description: Link not found.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
  /api/v1/analytics/{link_id}/geo:
    get:
      tags:
      - analytics
      summary: Get geo breakdown
      description: Return aggregated click counts grouped by country.
      operationId: get_geo_api_v1_analytics__link_id__geo_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
          title: Link Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakdownResponse'
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link.
        '404':
          description: Link not found.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
  /api/v1/analytics/{link_id}/devices:
    get:
      tags:
      - analytics
      summary: Get device breakdown
      description: Return aggregated click counts grouped by device, browser, and
        OS.
      operationId: get_devices_api_v1_analytics__link_id__devices_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
          title: Link Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceBreakdownResponse'
        '401':
          description: Authentication is required.
        '403':
          description: The current user cannot access this link.
        '404':
          description: Link not found.
        '422':
          description: Request validation failed.
        '429':
          description: Rate limit exceeded.
...
components:
  schemas:
    AnalyticsSummaryResponse:
      properties:
        link_id:
          type: integer
          title: Link Id
        total_clicks:
          type: integer
          title: Total Clicks
        unique_clicks:
          anyOf:
          - type: integer
          - type: 'null'
          title: Unique Clicks
        unique_clicks_note:
          type: string
          title: Unique Clicks Note
      type: object
      required:
      - link_id
      - total_clicks
      - unique_clicks
      - unique_clicks_note
      title: AnalyticsSummaryResponse
      description: Summary analytics for a single link.
      example:
        link_id: 742910234935296
        total_clicks: 1280
        unique_clicks_note: Unique clicks are not tracked yet; current schema stores
          only privacy-preserving event hashes and aggregate counts.
    BreakdownItem:
      properties:
        dimension:
          type: string
          title: Dimension
        clicks:
          type: integer
          title: Clicks
      type: object
      required:
      - dimension
      - clicks
      title: BreakdownItem
      description: One grouped analytics breakdown item.
    BreakdownResponse:
      properties:
        link_id:
          type: integer
          title: Link Id
        items:
          items:
            $ref: '#/components/schemas/BreakdownItem'
          type: array
          title: Items
      type: object
      required:
      - link_id
      - items
      title: BreakdownResponse
      description: Generic grouped analytics response.
      example:
        items:
        - clicks: 720
          dimension: US
        - clicks: 240
          dimension: IN
        link_id: 742910234935296
...
    DeviceBreakdownItem:
      properties:
        device_type:
          type: string
          title: Device Type
        browser:
          type: string
          title: Browser
        os:
          type: string
          title: Os
        clicks:
          type: integer
          title: Clicks
      type: object
      required:
      - device_type
      - browser
      - os
      - clicks
      title: DeviceBreakdownItem
      description: One device/browser/OS analytics breakdown item.
    DeviceBreakdownResponse:
      properties:
        link_id:
          type: integer
          title: Link Id
        items:
          items:
            $ref: '#/components/schemas/DeviceBreakdownItem'
          type: array
          title: Items
      type: object
      required:
      - link_id
      - items
      title: DeviceBreakdownResponse
      description: Device analytics response.
      example:
        items:
        - browser: Chrome
          clicks: 620
          device_type: desktop
          os: Windows
        link_id: 742910234935296
...
    LoginRequest:
      properties:
        email:
          type: string
          maxLength: 320
          minLength: 3
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: LoginRequest
      description: Request body for password login.
      example:
        email: ada@example.com
        password: correct-horse-battery
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          minLength: 1
          title: Refresh Token
      additionalProperties: false
      type: object
      required:
      - refresh_token
      title: RefreshRequest
      description: Request body for refresh-token rotation or logout.
      example:
        refresh_token: eyJhbGciOiJIUzI1NiJ9...
    RegisterRequest:
      properties:
        email:
          type: string
          maxLength: 320
          minLength: 3
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: RegisterRequest
      description: Request body for creating a password-backed user account.
      example:
        email: ada@example.com
        password: correct-horse-battery
    TimeseriesPoint:
      properties:
        date:
          type: string
          format: date
          title: Date
        clicks:
          type: integer
          title: Clicks
      type: object
      required:
      - date
      - clicks
      title: TimeseriesPoint
      description: One time-series analytics point.
    TimeseriesResponse:
      properties:
        link_id:
          type: integer
          title: Link Id
        granularity:
          type: string
          title: Granularity
        points:
          items:
            $ref: '#/components/schemas/TimeseriesPoint'
          type: array
          title: Points
      type: object
      required:
      - link_id
      - granularity
      - points
      title: TimeseriesResponse
      description: Time-series analytics response.
      example:
        granularity: day
        link_id: 742910234935296
        points:
        - clicks: 42
          date: '2026-06-14'
        - clicks: 58
          date: '2026-06-15'
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      title: TokenResponse
      description: Access and refresh JWT response.
      example:
        access_token: eyJhbGciOiJIUzI1NiJ9...
        expires_in: 900
        refresh_token: eyJhbGciOiJIUzI1NiJ9...
        token_type: bearer
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login